+++ lghorizon/exceptions.py
