--- lghorizon/models/__init__.py
@@ -1,12 +1,0 @@
-"""Models for LG Horizon."""
-
-from .lghorizon_auth import LGHorizonAuth
-from .lghorizon_config import LGHorizonServicesConfig
-from .lghorizon_message import LGHorizonMessage, LGHorizonStatusMessage
-
-__all__ = [
-    "LGHorizonAuth",
-    "LGHorizonServicesConfig",
-    "LGHorizonMessage",
-    "LGHorizonStatusMessage",
-]
